requirements.txt
matplotlib==3.7.2
pyserial==3.5